+++ CMakeLists.txt
@@ -1,7 +1,17 @@
 cmake_minimum_required(VERSION 3.28)
-project(ConversorUnidades C)
+project(ConversorDeUnidadesV2 C)
 
 set(CMAKE_C_STANDARD 11)
 
-add_executable(ConversorUnidades main.c
-        unidade.h)
+add_executable(ConversorDeUnidadesV2 main.c
+        unidades.c
+        unid.h
+        multunid.c
+        validarentrada.c
+        verifstring.c
+        dadossalvos.c
+        convunit.c
+        temperatura.c
+        funçãocategoria.c
+        salvarconv.c
+)
