+++ cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-f92daa908a00cfde3479.json
@@ -27,7 +27,7 @@
 					[
 						0
 					],
-					"name" : "ConversorUnidades",
+					"name" : "ConversorDeUnidadesV2",
 					"targetIndexes" : 
 					[
 						0
@@ -38,9 +38,9 @@
 			[
 				{
 					"directoryIndex" : 0,
-					"id" : "ConversorUnidades::@6890427a1f51a3e7e1df",
-					"jsonFile" : "target-ConversorUnidades-Debug-49fd5ea78eea7b7f31d2.json",
-					"name" : "ConversorUnidades",
+					"id" : "ConversorDeUnidadesV2::@6890427a1f51a3e7e1df",
+					"jsonFile" : "target-ConversorDeUnidadesV2-Debug-0a01e226d162c2e3dc2b.json",
+					"name" : "ConversorDeUnidadesV2",
 					"projectIndex" : 0
 				}
 			]
