+++ cmake-build-debug/CMakeCache.txt
@@ -84,7 +84,7 @@ CMAKE_FIND_PACKAGE_REDIRECTS_DIR:STATIC=C:/Users/jmrjr/CLionProjects/ConversorUn
 CMAKE_GNUtoMS:BOOL=OFF
 
 //Install path prefix, prepended onto install directories.
-CMAKE_INSTALL_PREFIX:PATH=C:/Program Files (x86)/ConversorUnidades
+CMAKE_INSTALL_PREFIX:PATH=C:/Program Files (x86)/ConversorDeUnidadesV2
 
 //Path to a program.
 CMAKE_LINKER:FILEPATH=C:/Program Files/JetBrains/CLion 2023.2/bin/mingw/bin/ld.exe
@@ -128,7 +128,7 @@ CMAKE_PROJECT_DESCRIPTION:STATIC=
 CMAKE_PROJECT_HOMEPAGE_URL:STATIC=
 
 //Value Computed by CMake
-CMAKE_PROJECT_NAME:STATIC=ConversorUnidades
+CMAKE_PROJECT_NAME:STATIC=ConversorDeUnidadesV2
 
 //Path to a program.
 CMAKE_RANLIB:FILEPATH=C:/Program Files/JetBrains/CLion 2023.2/bin/mingw/bin/ranlib.exe
@@ -214,13 +214,13 @@ CMAKE_TAPI:FILEPATH=CMAKE_TAPI-NOTFOUND
 CMAKE_VERBOSE_MAKEFILE:BOOL=FALSE
 
 //Value Computed by CMake
-ConversorUnidades_BINARY_DIR:STATIC=C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug
+ConversorDeUnidadesV2_BINARY_DIR:STATIC=C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug
 
 //Value Computed by CMake
-ConversorUnidades_IS_TOP_LEVEL:STATIC=ON
+ConversorDeUnidadesV2_IS_TOP_LEVEL:STATIC=ON
 
 //Value Computed by CMake
-ConversorUnidades_SOURCE_DIR:STATIC=C:/Users/jmrjr/CLionProjects/ConversorUnidades
+ConversorDeUnidadesV2_SOURCE_DIR:STATIC=C:/Users/jmrjr/CLionProjects/ConversorUnidades
 
 
 ########################
