+++ cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,3 @@
-C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug/CMakeFiles/ConversorUnidades.dir
+C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug/CMakeFiles/ConversorDeUnidadesV2.dir
 C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug/CMakeFiles/edit_cache.dir
 C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug/CMakeFiles/rebuild_cache.dir
