+++ cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,4 +1,10 @@
 "C:\Program Files\JetBrains\CLion 2023.2\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug "-DCMAKE_MAKE_PROGRAM=C:/Program Files/JetBrains/CLion 2023.2/bin/ninja/win/x64/ninja.exe" -G Ninja -S C:\Users\jmrjr\CLionProjects\ConversorUnidades -B C:\Users\jmrjr\CLionProjects\ConversorUnidades\cmake-build-debug
--- Configuring done (0.2s)
+-- The C compiler identification is GNU 13.1.0
+-- Detecting C compiler ABI info
+-- Detecting C compiler ABI info - done
+-- Check for working C compiler: C:/Program Files/JetBrains/CLion 2023.2/bin/mingw/bin/gcc.exe - skipped
+-- Detecting C compile features
+-- Detecting C compile features - done
+-- Configuring done (10.8s)
 -- Generating done (0.0s)
 -- Build files have been written to: C:/Users/jmrjr/CLionProjects/ConversorUnidades/cmake-build-debug
