+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: May 31 19:58 Hora oficial do Brasil
+Start testing: Jun 04 16:27 Hora oficial do Brasil
 ----------------------------------------------------------
-End testing: May 31 19:58 Hora oficial do Brasil
+End testing: Jun 04 16:27 Hora oficial do Brasil
