+++ cmake-build-debug/cmake_install.cmake
@@ -2,7 +2,7 @@
 
 # Set the install prefix
 if(NOT DEFINED CMAKE_INSTALL_PREFIX)
-  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/ConversorUnidades")
+  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/ConversorDeUnidadesV2")
 endif()
 string(REGEX REPLACE "/$" "" CMAKE_INSTALL_PREFIX "${CMAKE_INSTALL_PREFIX}")
