+++ convunit.c
@@ -1,0 +1,4 @@
+// Função para converter entre unidades
+double converter(double valor, double multiplicadorEntrada, double multiplicadorSaida) {
+    return valor * multiplicadorEntrada / multiplicadorSaida;
+}
