+++ unid.h
@@ -1,0 +1,4 @@
+typedef struct Unidade {
+    char nome[30];
+    double multiplicador; // fator de multiplicação para conversão para a unidade base
+} Unidade;
